Add modules create_rental and list_people
